requirements.txt
Django==4.2.13
djangorestframework==3.15.1
drf-yasg==1.21.7
gunicorn==25.3.0
whitenoise==6.6.0
PyJWT==2.10.1
django-cors-headers==4.3.1
setuptools